Add keychain bash script to handle ssh-agent

+++ install.sh
@@ -1,9 +1,5 @@
 #!/bin/bash
 
-RM_COLORS='\033[0m'
-GREEN='\033[0;32m'
-ORANGE='\033[0;33m'
-
 # Find current script dir
 SOURCE="${BASH_SOURCE[0]}"
 while [ -h "$SOURCE" ]; do # resolve $SOURCE until the file is no longer a symlink
@@ -12,6 +8,7 @@ while [ -h "$SOURCE" ]; do # resolve $SOURCE until the file is no longer a symli
     [[ $SOURCE != /* ]] && SOURCE="$DIR/$SOURCE" # if $SOURCE was a relative symlink, we need to resolve it relative to the path where the symlink file was located
 done
 DIR="$( cd -P "$( dirname "$SOURCE" )" && pwd )"
+BIN_DIR=$HOME/.local/bin
 
 # Read input with default value
 read -p "Install all files for '$USER' user ? (Y/n)" -n 1 choice
@@ -20,6 +17,10 @@ choice=${choice:-"y"};
 # Yes or no ?
 case ${choice} in
     y|Y)
+        if [ ! -d $BIN_DIR ]; then
+            mkdir -p $HOME/.local/bin
+        fi
+
         # Config files
         FILES=(
             .bashrc
@@ -32,7 +33,6 @@ case ${choice} in
             .tmux.conf
             .vim
             .vimrc
-            .zscript
             .zshrc
             .zshrc_ps1
         )
@@ -48,6 +48,28 @@ case ${choice} in
             else
                 ORIGINAL_FILE="${DIR}/${FILE}"
                 SYMBOLIC_LINK="${HOME}/${FILE}"
+                `ln -s "${ORIGINAL_FILE}" "${SYMBOLIC_LINK}"`
+                echo "${FILE} linked"
+            fi
+        done
+
+        # Binary files
+        FILES=(
+            bin/z.sh
+            bin/keychain
+        )
+
+        for FILE in  ${FILES[@]}
+        do
+            if [ -f ~/${FILE} ]
+            then
+                echo "File exists : ${FILE} not linked"
+            elif [ -d ~/${FILE} ]
+            then
+                echo "Folder exists: ${FILE} not linked"
+            else
+                ORIGINAL_FILE="${DIR}/${FILE}"
+                SYMBOLIC_LINK="${HOME}/.local/${FILE}"
                 `ln -s "${ORIGINAL_FILE}" "${SYMBOLIC_LINK}"`
                 echo "${FILE} linked"
             fi
